feat!: delete old e2e
